# Simplifies shouldComponentUpdate (open discussion)