Simplifications to `shouldComponentUpdate` now as we have `_.isEqual`. 

Started implementation based on some comments from #76 - and this is a continuation of the discussion. Not yet implemented "rebase" support. As that may be a separate change.

 Nor is comparable implemented. As there are some discussion if we need to do deep equal on immutable structures that has been said to not be equal through reference checks.