@torgeir 2015-02-28
Faster/easier should component update :+1: ! 

Though, I think for something like this we need some hard numbers on comparison speeds before and after these changes.