@mikaelbr 2015-02-28
I'm working on a benchmark. Copied Immutable.js' benchmark and trying to run it now. It compares previous dist file with current uncommitted dist.