@Dashed 2015-02-28
:+1: Awesome! `shouldComponentUpdate` looks so much easier to read, extend, and refactor.

@mikaelbr Looking forward to see the benchmarks.

I wonder if `isEqualProps` and `isEqualState` can be made more DRY. 

-----

@Gozala Do you agree that there's a compromise where one can override `isEqualState` and/or `isEqualProps` for a more aggressive `shouldComponentUpdate`?