@mikaelbr 2015-02-28
Thanks for good feedback.

I haven't gotten the benchmark to work as I want yet, and haven't had the time to work on it too much, but it'll get there.