@mikaelbr 2015-03-01
Here's the initial relative benchmarks for the current dist and this dist:

```
shouldComponentUpdate on no input
  Old:   180,556   182,225   183,925 ops/sec
  New:   255,906   258,463   261,071 ops/sec
  compare: 1 1
  diff: 41.83%
  rme: 0.96%
shouldComponentUpdate when immutables are same
  Old:    71,202    71,945    72,703 ops/sec
  New:   107,211   108,415   109,646 ops/sec
  compare: 1 1
  diff: 50.69%
  rme: 1.08%
shouldComponentUpdate when many immutables are passed as the same reference
  Old:    10,573    10,878    11,201 ops/sec
  New:    11,915    12,060    12,208 ops/sec
  compare: 1 1
  diff: 10.86%
  rme: 2.21%
shouldComponentUpdate when large object is passed
  Old:    20,329    20,540    20,754 ops/sec
  New:    22,322    22,502    22,685 ops/sec
  compare: 1 1
  diff: 9.55%
  rme: 0.92%
shouldComponentUpdate when two large object is passed
  Old:     1,265     1,280     1,296 ops/sec
  New:     1,523     1,537     1,552 ops/sec
  compare: 1 1
  diff: 20.07%
  rme: 1.06%
shouldComponentUpdate when two large object is passed but as same input
  Old:     2,381     2,424     2,469 ops/sec
  New:    29,189    29,421    29,656 ops/sec
  compare: 1 1
  diff: 1113.38%
  rme: 1.39%
shouldComponentUpdate when immutables are different
  Old:    40,521    42,657    45,029 ops/sec
  New:    42,122    45,864    50,336 ops/sec
  compare: 1 1
  diff: 7.51%
  rme: 7.3%
shouldComponentUpdate when long arrays (1000 elements) are the same
  Old:     1,587     1,623     1,660 ops/sec
  New:     8,076     8,466     8,895 ops/sec
  compare: 1 1
  diff: 421.57%
  rme: 3.76%
shouldComponentUpdate when long arrays (1000 elements) are the same reference
  Old:    13,032    13,229    13,432 ops/sec
  New:    13,451    13,720    14,000 ops/sec
  compare: 1 1
  diff: 3.71%
  rme: 1.77%
shouldComponentUpdate when same input is passed (long array, 1000 items)
  Old:     6,381     6,461     6,543 ops/sec
  New:     8,359     9,117    10,027 ops/sec
  compare: 1 1
  diff: 41.1%
  rme: 6.47%
all done
```

Haven't looked too much into this, so these numbers might be wrong, but they seem reasonable. I didn't get cursors to work, but immutable structures should works just as good.

The tests can be found here: https://gist.github.com/mikaelbr/b4760464013fe636186f
Again, I haven't looked too much into this and mostly just copy/pasted and tested it out - so there's very likely that there are some dumb tests.