@Dashed 2015-03-01
Here are my results with `iojs-v1.4.1`.

```
shouldComponentUpdate on no input
  ...running...  
  Old:   134,954   137,313   139,756 ops/sec
  New:   131,645   136,062   140,785 ops/sec
  compare: 1 1
  diff: -0.92%
  rme: 2.67%
shouldComponentUpdate when immutables are same
  ...running...  
  Old:    46,624    51,573    57,696 ops/sec
  New:    55,795    65,344    78,838 ops/sec
  compare: 1 1
  diff: 26.7%
  rme: 14.24%
shouldComponentUpdate when many immutables are passed as the same reference
  ...running...  
  Old:     7,545     7,653     7,764 ops/sec
  New:     8,099     8,217     8,339 ops/sec
  compare: 1 1
  diff: 7.37%
  rme: 1.44%
shouldComponentUpdate when large object is passed
  ...running...  
  Old:    10,774    10,864    10,956 ops/sec
  New:    11,476    11,635    11,800 ops/sec
  compare: 1 1
  diff: 7.09%
  rme: 1.14%
shouldComponentUpdate when two large object is passed
  ...running...  
  Old:       649       662       676 ops/sec
  New:       838       852       867 ops/sec
  compare: 1 1
  diff: 28.71%
  rme: 1.89%
shouldComponentUpdate when two large object is passed but as same input
  ...running...  
  Old:     1,134     1,149     1,165 ops/sec
  New:    14,020    14,176    14,334 ops/sec
  compare: 1 1
  diff: 1132.72%
  rme: 1.22%
shouldComponentUpdate when immutables are different
  ...running...  
  Old:    27,957    28,465    28,991 ops/sec
  New:    40,528    41,205    41,906 ops/sec
  compare: 1 1
  diff: 44.75%
  rme: 1.74%
shouldComponentUpdate when long arrays (1000 elements) are the same
  ...running...  
  Old:       585       593       602 ops/sec
  New:     2,243     2,285     2,329 ops/sec
  compare: 1 1
  diff: 285.03%
  rme: 1.65%
shouldComponentUpdate when long arrays (1000 elements) are the same reference
  ...running...  
  Old:     4,664     4,749     4,837 ops/sec
  New:     4,758     4,849     4,943 ops/sec
  compare: 1 1
  diff: 2.11%
  rme: 1.86%
shouldComponentUpdate when same input is passed (long array, 1000 items)
  ...running...  
  Old:     4,529     4,628     4,732 ops/sec
  New:     5,084     5,179     5,278 ops/sec
  compare: 1 1
  diff: 11.89%
  rme: 2.04%
all done
```